Moved DashboardController and DashboardControllerIT to package interfaces/web/application.
