+++ abixen-platform-core/src/test/java/com/abixen/platform/core/interfaces/CommentVoteControllerTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.abixen.platform.core.controller;
+package com.abixen.platform.core.interfaces;
 
 import com.abixen.platform.common.application.dto.FormErrorDto;
 import com.abixen.platform.common.domain.model.enumtype.CommentVoteType;
